requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Pillow==10.0.0
PyPDF2==3.0.0
behave==1.3.3